fix: moved xcm utils to prevent errors in production
