+++ client/packages/sdk/src/utils/index.ts
@@ -1,4 +1,3 @@
 import * as Substrate from "./substrate"
-import { generateXcmTransferParameters } from "./xcm"
 
-export { Substrate, generateXcmTransferParameters }
+export { Substrate }
